+++ source/helper/stringable.js
@@ -1,9 +1,0 @@
-const type = require('./type.js');
-
-/**
- * Test whether given input is or can be casted to a string
- *
- * @param {*} input
- * @returns {boolean} stringable
- */
-module.exports = (input) => /string|object/.test(type(input));